[committed] Fix compromised ARC test
Jakub's recent changes for pr120231 compromised this arc port specific test.
Essentially we collapse the entire FMA sequence down to a constant and thus
never emit the FMA instruction the test wants to see.

If we make "a" an extern so that we don't know its value the optimizers can't
collapse the calculation away completely and we can verify that we get an FMA
operation.

Pushing to the trunk.

gcc/testsuite
	* gcc.target/arc/fma-1.c: Make "a" extern so the optimizers can
	see any value and optimize away the key computation.