cselim: change how to detect no load/stores after store in single_trailing_store_in_bb
So the current code iterates through the uses of the vdef of the store to see if there
is an use of the vdef inside the bb. Except in this case we know the only usage will
be a phi node, so change the loop into calling single_imm_use and then checking if the
usage statement was the PHI node.

Bootstrapped and tested on x86_64-linux-gnu.

Also did a quick test by adding a return false after the checks of single_trailing_store_in_bb
in cond_if_else_store_replacement to make sure this code still does the correct thing.

gcc/ChangeLog:

	* tree-ssa-phiopt.cc (single_trailing_store_in_bb): Add vphi argument.
	Check for single use of the vdef of the store instead of a loop
	and check vdef's single use statement is the same as vphi.
	(cond_if_else_store_replacement): Update call to single_trailing_store_in_bb.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>