ada: Fix constraint-related legality checks in extended return statements
Legality checks in extended return statements were (almost) literally
implementing the RM rules, but the when analyzing the return object declaration
we replace the nominal subtype of that object with its constrained subtype.
(It is a bit odd to have such an expansion activity in analysis, but we already
rely on this particular expansion in quite a few places).

gcc/ada/ChangeLog:

	* sem_ch3.adb (Check_Return_Subtype_Indication): Use the nominal
	subtype of a return object; literally implement the RM rule about
	elementary types; check for static subtype compatibility both when
	the subtype is given as a subtype mark and a subtype indication.