More API for IPA profile manipulation
This patch attempts to make IPA profile manipulation easier.  It introduces

 node->scale_profile_to (count)
    which can be used to scale profile to a given IPA count.
    If IPA count is zero, then local profile is preserved and proper variant
    of global0 count is used.

 node->make_profile_local
    this can be used to drop IPA profile but keep local profile

 node->make_profile_global0
    this can be used to make IPA profile 0 but keep local
    profile.

Most of this can be accomplished by existing apply_scale.  I.e.
 - node->scale_profile_to (count)
     corresponds to
    node->apply_scale (count, node->count),
 - node->make_profile_local
     corresponds to
   node->apply_scale (node->count.guessed_local (), node->count)

I think new API is more clean about what intention is and less error prone.
Also it handles some side cases when entry block of profile happens to be 0,
but body is non-zero (by profile inconsistencies).  In this case the scaling
API did kind of random things.

I noticed three bugs in ipa-cp (two already in released GCCs while one mine
introduced by last patch):

@@ -4528,7 +4528,7 @@ lenient_count_portion_handling (profile_count remainder, cgraph_node *orig_node)
   if (remainder.ipa_p () && !remainder.ipa ().nonzero_p ()
       && orig_node->count.ipa_p () && orig_node->count.ipa ().nonzero_p ()
       && opt_for_fn (orig_node->decl, flag_profile_partial_training))
-    remainder = remainder.guessed_local ();
+    remainder = orig_node->count.guessed_local ();

The code was intended to drop IPA profile to local when remainder is 0.
In this case orig_node->count is some non-zero count but all of control flow
was redirected to a clone which means that remainer is 0 (adjusted).
Doing

     remainder = remainder.guessed_local ();

will turn it into 0 (guessed_local) and the scalling will then multiply
all counts by 0 and turn them tinto guessed local.
We want to keep original count but reduce the quality. i.e.

     remainder = orig_node->count.guessed_local ();

Second problem is:
       /* TODO: Profile has alreay gone astray, keep what we have but lower it
 	 to global0 category.  */
       remainder = orig_node->count.global0 ();
global0 means that converting to ipa count will be precise 0. Since we lost track
it should be adjusted 0 :)

Finally in:
  new_sum = orig_node_count.combine_with_ipa_count (new_sum);
  orig_node->count = remainder;

  new_node->apply_scale (new_sum, new_node->count);

   if (!orig_edges_processed)
    orig_node->apply_scale (remainder, orig_node->count);
    orig_node->scale_profile_to (remainder);

orig_node->count is first set to remainder and then scalling is done
(which in turn does nothing).

This is bug I introduced in last path which should have removed
orig_node->count = remainder.  As a result now counts of cgraph edges are not
adjusted correctly.  I am sorry for that.

gcc/ChangeLog:

	* cgraph.cc (cgraph_node::make_profile_local): New member function.
	(cgraph_node::make_profile_global0): New member function.
	(cgraph_node::apply_scale): Do not call adjust_for_ipa_scalling.
	(cgraph_node::scale_profile_to): New member function.
	* cgraph.h (cgraph_node::make_profile_local,
	cgraph_node::make_profile_global0, cgraph_node::scale_profile_to):
	Declare.
	* ipa-cp.cc (lenient_count_portion_handling): Fix logic dropping count
	to local.
	(update_counts_for_self_gen_clones): Use scale_profile_to.
	(update_profiling_info): Use make_profile_local, make_profile_global0
	and scale_profile_to.
	(update_specialized_profile): Likewise.
	* ipa-inline-transform.cc (clone_inlined_nodes): Call
	adjust_for_ipa_scalling.