c++: add -Wsfinae-incomplete
We already error about a type or function definition causing a concept check
to change value, but it would be useful to diagnose this for other SFINAE
contexts as well; the memoization problem also affects templates.  So
-Wsfinae-incomplete remembers if we've failed a requirement for a complete
type/deduced return type in a non-tf_error context, and later warns if the
type/function becomes complete.

This warning is enabled by default; I think the signal-to-noise ratio is
high enough to warrant that, and it catches things that are likely to make
the program "ill-formed, no diagnostic required".

friend87.C is an interesting case; this could be considered a false positive
because it is using friend injection to define the auto function to
implement a compile-time counter.  I think this is sufficiently pathological
that it's fine to expect people who want to play this sort of game to
suppress the warning.

The data for this warning uses GTY((cache)) to persist through GC, but allow
entries to be discarded if the key is not otherwise marked.

I don't think it's desirable to export/import this information in modules,
it makes sense for it to be local to a single TU.

-Wsfinae-incomplete=2 adds a warning at the point of failure, which is
primarily intended to help with debugging warnings from the default mode.

gcc/ChangeLog:

	* doc/invoke.texi: Document -Wsfinae-incomplete.

gcc/c-family/ChangeLog:

	* c.opt: Add -Wsfinae-incomplete.
	* c.opt.urls: Regenerate.

gcc/cp/ChangeLog:

	* constraint.cc (failed_completions_map): New.
	(note_failed_type_completion): Rename from
	note_failed_type_completion_for_satisfaction.  Add
	-Wsfinae-incomplete handling.
	(failed_completion_location): New.
	* class.cc (finish_struct_1): Add -Wsfinae-incomplete warning.
	* decl.cc (require_deduced_type): Adjust.
	(finish_function): Add -Wsfinae-incomplete warning.
	* typeck.cc (complete_type_or_maybe_complain): Adjust.
	(cxx_sizeof_or_alignof_type): Call note_failed_type_completion.
	* pt.cc (dependent_template_arg_p): No longer static.
	* cp-tree.h: Adjust.

libstdc++-v3/ChangeLog:

	* testsuite/20_util/is_complete_or_unbounded/memoization.cc
	* testsuite/20_util/is_complete_or_unbounded/memoization_neg.cc:
	Expect -Wsfinae-incomplete.

gcc/testsuite/ChangeLog:

	* g++.dg/template/friend87.C
	* g++.dg/cpp2a/concepts-complete1.C
	* g++.dg/cpp2a/concepts-complete2.C
	* g++.dg/cpp2a/concepts-complete3.C
	* g++.dg/cpp2a/concepts-complete4.C: Expect -Wsfinae-incomplete.