amdgcn: add more insn patterns using vec_duplicate
These new insns allow more efficient use of scalar inputs to 64-bit vector
add and mul.  Also, the patch adjusts the existing mul.._dup because it was
actually a dup2 (the vec_duplicate is on the second input), and that was
inconveniently inconsistent.

The patterns are generally useful, but will be used directly by a follow-up
patch.

gcc/ChangeLog:

	* config/gcn/gcn-valu.md (add<mode>3_dup): New.
	(add<mode>3_dup_exec): New.
	(<su>mul<mode>3_highpart_dup<exec>): New.
	(mul<mode>3_dup): Move the vec_duplicate to operand 1.
	(mul<mode>3_dup_exec): New.
	(vec_series<mode>): Adjust call to gen_mul<mode>3_dup.
	* config/gcn/gcn.cc (gcn_expand_vector_init): Likewise.